Week6/Day5/Exercises/bike_store/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
Faker==19.2.0
